implemented swalalert2

+++ src/app/components/cart-app.component.ts
@@ -5,6 +5,7 @@ import { CartItem } from '../models/cartitem';
 import { NavbarComponent } from './navbar/navbar.component';
 import { Router, RouterOutlet } from '@angular/router';
 import { SharingDataService } from '../services/sharing-data.service';
+import Swal from 'sweetalert2'
 
 @Component({
   selector: 'cart-app',
@@ -55,24 +56,49 @@ export class CartAppComponent implements OnInit {
     this.router.navigate([['/cart',{
       state: { items: this.items, total: this.total }
     }]]);
+    Swal.fire({
+      title: 'Shopping',
+      text: 'Nuevo producto agregado',
+      icon: 'success',
     })
-
+    });
   }
 
   onDeleteCart(){
     this.sharingDataService.idProductEventEmmiter.subscribe(id => {
-      this.items = this.items.filter(item=>item.product.id !== id);
-      if(this.items.length == 0){
-        sessionStorage.removeItem('cart');
-        sessionStorage.clear();
-      }
-      this.calculateTotal();
-      this.saveSession();
-      this.router.navigateByUrl('/',{skipLocationChange:true}).then(()=>{
-        this.router.navigate(['/cart'],{
-          state:{items: this.items, total: this.total}
-        })
-      })
+      Swal.fire({
+        title: "Desea eliminar?",
+        text: "Item se eliminara del carrito",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, delete it!"
+      }).then((result) => {
+        if (result.isConfirmed) {
+          
+          this.items = this.items.filter(item=>item.product.id !== id);
+          if(this.items.length == 0){
+            sessionStorage.removeItem('cart');
+            sessionStorage.clear();
+          }
+
+          this.calculateTotal();
+          this.saveSession();
+          this.router.navigateByUrl('/',{skipLocationChange:true}).then(()=>{
+            this.router.navigate(['/cart'],{
+              state:{items: this.items, total: this.total}
+            })
+          })
+
+          Swal.fire({
+            title: "Eliminado!",
+            text: "Se ha eliminado",
+            icon: "success"
+          });
+        }
+      });
+
     });
   }
